+++ src/processx.h
@@ -37,7 +37,7 @@ extern "C" {
 /* API from R */
 
 SEXP processx_exec(SEXP command, SEXP args,
-		   SEXP std_in,  SEXP std_out, SEXP std_err,
+		   SEXP std_in,  SEXP std_out, SEXP std_err, SEXP env,
 		   SEXP windows_verbatim_args,
 		   SEXP windows_hide_window, SEXP private_, SEXP cleanup,
 		   SEXP wd, SEXP encoding);